Lift tests that only test public API into tests/
There are some tests, especially the parse ones, that stay internal
because they access the internals of the parsed objects.
